Added Features
Added a win-loss ratio. The ratio is displayed in green if it is greater than or equal to 1 and red if it is lesser than 1.
Updated the goal difference to be displayed in green if it is greater than or equal to 0 and red if lesser than 0.
Updated the trend to only display the top 10 matches of the history.